d/initramfs/cryptroot-script: Allow some retries on local-block stage
Currently, cryptsetup try mounting encrypted rootfs only in local-top,
failing if not possible. Hence cases like LUKS on top of RAID1 (if the
array is degraded) cannot work properly.

This patch allows cryptsetup to retry on local-block stage, relying in
a heuristic based on ROOTDELAY and on initramfs looping at local-block
phase. We also added a script to local-bottom stage in order to clean
some control files used to track progress in local-block.

The tests with Debian Buster were successful; we created RAID1 and a
LUKS volume on top of it, to hold the rootfs. By removing one disk of
RAID1, we fail to boot without the patch (initramfs-tools drops into a
shell), whereas with the patch we have the boot succeeding.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@canonical.com>