Move SSH token plugin stuff into new binary package 'cryptsetup-ssh'.
That plugin is arguably not useful for everyone and we can save the
‘Depends: libssh-4’ on cryptsetup-bin by moving cryptsetup-ssh(8) and
libcryptsetup-token-ssh.so to a separate package.

Moreover in upstream's own word this is an experimental plugin so it
probably doesn't hurt to emphasize that in the package description.

Since no Debian release supports LUKS2 SSH tokens (it was added in
2:2.4.0~rc0-1+exp1), and since they are currently in experimental stage,
we don't let cryptsetup-bin or cryptsetup depend on the new binary
package.  Users who need that feature will need to install it manually.