d/t: Add wrapper script for local cryptroot-* tests.
Running autopkgtest(1) causes a significant overhead while developing
initramfs scripts, and it makes sense to have the possibility to fire off
VMs locally and run all (or some) cryptroot-* tests.  This change makes
it possible to design custom-* tests which might be not worth running as
autopkgtests.

We create a temporary APT repository with the local .deb and pass it
along to d/t/utils/cryptroot-common so the right version is made
available to the guest.