d/t/cryptroot-*: Communicate with guests on /dev/hvc0.
Instead of the serial port /dev/ttyS0.  But keep kernel messages to
/dev/ttyS0 only (and leave /dev/console there too).  That way
unsolicited kernel log messages don't mess up with command prompt or
output.

The cryptsetup prompt at initramfs stage remains on /dev/console.

We previously exposed /dev/hvc0 as a side channel to interact with the
guest, but it was never used so we now keep it internal.