d/t/cryptroot-*: Use the QMP "quit" command to destroy guests early.
(Except for cryptroot-sysvinit where we need to test our init scripts.)
There is no need to run through the shutdown phase when the test has
passed already.

To do this we add QMP support to the communication logic (using the
monitor channel).  This requires the addition of a new dependency
‘libjson-perl’ for autopkgtests.