d/t/cryptroot-*: Manually create merged-/usr layout and install usr-is-merged.
Rather than relying on the ‘usrmerge’ package to convert the layout, as
unpacking ‘usrmerge’ fails when /usr is on a separate file system:

    Preparing to unpack .../usrmerge_30+nmu1_all.deb ...
    /usr is a standalone filesystem, this requires using an initramfs.
    dpkg: error processing archive /media/__essential__/usrmerge_30+nmu1_all.deb (--install):
     new usrmerge package pre-installation script subprocess returned error exit status 1

The merged-/usr layout is controlled by a new ‘MERGED_USR=[yes|no]’
option, which default to ‘no’ up to Bullseye, and to ‘yes’ since
Bookworm.

Note that init-system-helpers 1.65.2 (from current sid) has ‘Depends:
usrmerge | usr-is-merged’, so setting MERGED_USR=no there actually
*does* setup a merged-/usr layout (it's not created by the debootstrap
script but by the ‘usrmerge’ package).